Moved util function
